Create README.md
# Parking_App
This is an application that if you wish, stores the position of your parked vehicle, so it can be found easily.
From any location, you can reach your vehicle's location with the guidance of google maps. 
The app also includes features like parking statistics and recommends parking positions, 
calculated from places visited in the past.